+++ tests/TestCase/Utility/CheckupTest.php
@@ -53,7 +53,6 @@ class CheckupTest extends TestCase
     public function testApache()
     {
         $this->assertInstanceof(Apache::class, $this->Checkup->Apache);
-        $this->assertEquals(['modules', 'getVersion'], get_class_methods($this->Checkup->Apache));
         $this->assertArrayKeysEqual(['expires', 'rewrite'], $this->Checkup->Apache->modules());
         $this->assertRegExp('/^[\d\.]+$/', $this->Checkup->Apache->getVersion());
     }
@@ -65,7 +64,6 @@ class CheckupTest extends TestCase
     public function testBackups()
     {
         $this->assertInstanceof(Backups::class, $this->Checkup->Backups);
-        $this->assertEquals(['isWriteable'], get_class_methods($this->Checkup->Backups));
         $this->assertEquals([getConfig('DatabaseBackup.target') => true], $this->Checkup->Backups->isWriteable());
     }
 
@@ -76,7 +74,6 @@ class CheckupTest extends TestCase
     public function testElFinder()
     {
         $this->assertInstanceof(ElFinder::class, $this->Checkup->ElFinder);
-        $this->assertEquals(['isAvailable', 'getVersion'], get_class_methods($this->Checkup->ElFinder));
         $this->assertTrue($this->Checkup->ElFinder->isAvailable());
         $this->assertRegExp('/^[\d\.]+$/', $this->Checkup->ElFinder->getVersion());
     }
@@ -88,7 +85,6 @@ class CheckupTest extends TestCase
     public function testPHP()
     {
         $this->assertInstanceof(PHP::class, $this->Checkup->PHP);
-        $this->assertEquals(['extensions', 'getVersion'], get_class_methods($this->Checkup->PHP));
         $this->assertNotEmpty($this->Checkup->PHP->extensions());
         $this->assertRegExp('/^7\.\d+\.\d+$/', $this->Checkup->PHP->getVersion());
     }
@@ -100,7 +96,6 @@ class CheckupTest extends TestCase
     public function testPlugin()
     {
         $this->assertInstanceof(Plugin::class, $this->Checkup->Plugin);
-        $this->assertEquals(['getVersions'], get_class_methods($this->Checkup->Plugin));
         $this->assertArrayKeysEqual(['me_cms', 'others'], $this->Checkup->Plugin->getVersions());
         $this->assertNotEmpty($this->Checkup->Plugin->getVersions()['me_cms']);
         $this->assertNotEmpty($this->Checkup->Plugin->getVersions()['others']);
@@ -113,8 +108,6 @@ class CheckupTest extends TestCase
     public function testTMP()
     {
         $this->assertInstanceof(TMP::class, $this->Checkup->TMP);
-        $this->assertEquals(['isWriteable'], get_class_methods($this->Checkup->TMP));
-
         $result = $this->Checkup->TMP->isWriteable();
         $this->assertNotEmpty($result);
         foreach ($result as $path => $isWriteable) {
@@ -130,8 +123,6 @@ class CheckupTest extends TestCase
     public function testWebroot()
     {
         $this->assertInstanceof(Webroot::class, $this->Checkup->Webroot);
-        $this->assertEquals(['isWriteable'], get_class_methods($this->Checkup->Webroot));
-
         $result = $this->Checkup->Webroot->isWriteable();
         $this->assertNotEmpty($result);
         foreach ($result as $path => $isWriteable) {